% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-mixed-model.R
\name{diagnose_lmm}
\alias{diagnose_lmm}
\title{Diagnose a Linear Mixed Model}
\usage{
diagnose_lmm(lmm)
}
\arguments{
\item{lmm}{A fitted `lmer` model.}
}
\value{
A Shiny app object.
}
\description{
This function launches a Shiny dashboard to inspect diagnostics of an LMM.
}
